# openapi.json
{
  "components": {
    "schemas": {
      "FieldError": {
        "description": "``errors[]`` 的元素——422 欄位級明細（09 附錄 A：VALIDATION_FAILED）。",
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "LoginIn": {
        "description": "登入請求。\n\n``tenant_slug`` 是必填的，因為登入發生在租戶身分存在**之前**：RLS 之下，\n沒有租戶就查不到任何使用者，而 email 只在租戶內唯一（05 §3.1）。\nslug → tenant_id 由不含 RLS 的目錄表解析（apps/identity/models.py 的\n`TenantDirectory` 有完整說明）。\n\n未來改成子網域登入時，這個欄位改由 Host 標頭推導，契約可維持相容。",
        "properties": {
          "email": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "tenant_slug": {
            "maxLength": 63,
            "minLength": 1,
            "title": "Tenant Slug",
            "type": "string"
          }
        },
        "required": [
          "tenant_slug",
          "email",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
      "PasswordChangeIn": {
        "properties": {
          "current_password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "maxLength": 1024,
            "minLength": 8,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChangeIn",
        "type": "object"
      },
      "ProblemDetail": {
        "description": "RFC 9457 §3.1 的成員 + 本專案的 extension member（§3.2 允許）。",
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "09 附錄 A 的穩定錯誤碼；client 唯一該用來分支的值。無對應碼時省略",
            "title": "Code"
          },
          "detail": {
            "description": "這一次發生的情況；5xx 一律通用敘述，不含內部細節",
            "title": "Detail",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "4xx 的補充資訊",
            "title": "Details"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "422 的欄位級明細",
            "title": "Errors"
          },
          "request_id": {
            "description": "服務端產生的追蹤 id，與 X-Request-Id 標頭相同",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "description": "HTTP 狀態碼，與回應行一致",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "錯誤類型的人類可讀摘要；同一 type 恆定",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "錯誤類型的 URI reference；無額外語意時為 about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "request_id"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "TokenPairOut": {
        "description": "**只有 access token**。refresh 走 httpOnly cookie，不進 body。\n\n放進 body 的話前端就得自己存，而 localStorage 是 XSS 的直接目標；httpOnly\n讓 JavaScript 完全讀不到它。",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "TokenPairOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Lumina spike (ADR-001 bridge)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "auth_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "請求格式錯誤"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "不存在或無權可見"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "語意驗證失敗"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "內部錯誤"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "auth_logout",
        "parameters": [
          {
            "in": "cookie",
            "name": "refresh_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "請求格式錯誤"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "不存在或無權可見"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "語意驗證失敗"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "內部錯誤"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password/change": {
      "post": {
        "description": "改密碼並撤銷**全部**session（10 §2.1）。\n\n改密碼的典型情境就是「我懷疑帳號被盜」——只更新雜湊而不撤銷的話，攻擊者\n手上的 token 完全不受影響，而使用者以為自己已經處理完了。",
        "operationId": "auth_change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "請求格式錯誤"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "不存在或無權可見"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "語意驗證失敗"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "內部錯誤"
          }
        },
        "summary": "Change Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "auth_refresh",
        "parameters": [
          {
            "in": "cookie",
            "name": "refresh_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "請求格式錯誤"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "不存在或無權可見"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "語意驗證失敗"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "內部錯誤"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}